chore(fflonk): cleanup
